refactor: 테이블 업데이트 중간 연산 수정
