+++ src/main/java/com/example/masterplanbbe/application/batch/step/FirstStepWriter.java
@@ -1,7 +1,7 @@
 package com.example.masterplanbbe.application.batch.step;
 
 import com.example.masterplanbbe.application.batch.dto.IntermediateStepWriteDTO;
-import com.example.masterplanbbe.infrastructure.repository.BatchFirstStepJdbcRepository;
+import com.example.masterplanbbe.infrastructure.repository.BatchIntermediateStepJdbcRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.item.Chunk;
@@ -13,11 +13,11 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class FirstStepWriter implements ItemWriter<IntermediateStepWriteDTO> {
 
-    private final BatchFirstStepJdbcRepository batchFirstStepJdbcRepository;
+    private final BatchIntermediateStepJdbcRepository batchIntermediateStepJdbcRepository;
 
     @Override
     public void write(Chunk<? extends IntermediateStepWriteDTO> chunk) {
 //        log.info("FirstStepWriter 데이터 : {}", chunk.getItems());
-        batchFirstStepJdbcRepository.batchSave(chunk.getItems());
+        batchIntermediateStepJdbcRepository.batchSave(chunk.getItems());
     }
 }
